requirements.txt
Django==1.8.5
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
django-model-utils==2.3.1
factory-boy==2.6.0
fake-factory==0.5.3
html5lib==0.999
py==1.4.30
pytest==2.8.2
pytest-django==2.9.1
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2
fabric
django-http-proxy
